+++ src/commands/giveaway.ts
@@ -1,6 +1,6 @@
 import { Message, RichEmbed, MessageReaction, User, Collection } from 'discord.js';
 import Command from '../interfaces/command';
-import parseTime from '../utils/timeParser';
+import Time from '../utils/timeParser';
 import bot from '..';
 
 export default class GiveawayCommand implements Command {
@@ -10,32 +10,30 @@ export default class GiveawayCommand implements Command {
         usage: '&giveaway (time) (topic..)'
     }
 
-    reactions: string[] = ['🍞'];
-
     async run(message: Message, args: string[], messages: any): Promise<any> {
         if(args.length < 2)
             return message.reply(`${messages.use} ${this.info.usage}`);
 
-        let time: number | false = parseTime(args.shift());
-        if(!time)
+        let time: Time = new Time(args.shift());
+        if(time.invalid)
             return message.reply(`${messages.use} ${this.info.usage}`);
 
-        let reaction: string
-            = this.reactions[Math.floor(Math.random() * this.reactions.length)];
+        let reaction: string = '🍞';
 
-        let giveaway: Message | Message[] = await message.channel.send(
+        let giveaway: Message = await message.channel.send(
             new RichEmbed()
                 .setTitle('GIVEAWAY')
                 .setColor('RANDOM')
-                .setDescription(`${args.join(' ')}\n${messages.reactGiveaway.replace('{}', reaction)}`)
-        );
-        if(!(giveaway instanceof Message)) return;
-        
+                .setDescription(`
+                ${args.join(' ')}
+                ${messages.reactGiveaway.replace('{}', reaction)}
+                ${messages.endsIn} ${time.raw}
+                `)
+        ) as Message;
+
         giveaway.react(reaction);
 
         setTimeout((): void => {
-            if(!(giveaway instanceof Message)) return;
-
             giveaway.reactions
                 .find((r: MessageReaction): boolean => {
                     return r.emoji.toString() == reaction;
@@ -59,6 +57,6 @@ export default class GiveawayCommand implements Command {
                     
                     message.channel.send(`${winner.toString()} ${messages.won.replace('{}', args.join(' '))}`);
                 });
-        }, time);
+        }, time.ms);
     }
 }