+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 # 2.5 - (something)
+## 2.5.1
+- Fixed `permissions` command
+- Added `greeting` command
+- Optimization
+- *oof* Updated README version from `2.0.2` to `2.5.1`
+- Updated the time parser
+
 ## 2.5.0 (btw, this is probably the biggest update ever)
 - Fixed `rank, ranking, math` commands
 - Added `profile, exec, stats, language, prefix, stop, autorole` commands